requirements.txt
sqlalchemy
fastapi
fastapi[standard]
uvicorn
pyjwt
bcrypt==3.2.2
python-jose